requirements.txt
# external requirements
click
Sphinx
coverage
awscli>=1.29.35
flake8
python-dotenv>=0.5.1
requests
requests-mock
scipy
shapely
rasterio
xarray
pyproj
matplotlib
fastapi
hypercorn[trio]==0.14.3
zarr
pydantic
prometheus-client
python-multipart
PyJWT
cryptography
geosyspy==0.2.1
byoa